cleanup: move step and sidecar validation
Move step and sidecar validation functions with their call chain to
`container_validation.go` and fix small typos.

Create `validateSidecar` and move to `container_validation.go` to be
consistent with the `validateStep` function location. Inline
`validateSidecarName` function into `validateSidecar` to keep
consistency because the fields `sc.Image` and `sc.Script` are validated
in the caller.

Keep `validationSteps` and  `validateSidecars` in `task_validation.go`.

Issue #7442.
